Update 001a-imprint.md
Hi Anita, sorry for the delay in starting the translation, the last few weeks have been very busy.
I started now and I'm forwarding the translation proposals.

Any questions don't hesitate to ask and if I'm doing something wrong with the files let me know and i'll adjust here.

The title of the book when translated into Portuguese has a negative connotation: "(A)prenda Bitcoin".

Something like:
"(A)arrest Bitcoin."

Because the word "prenda" means to put in jail. That's why we can discuss together in the end possibilities of keeping the pun that you made in English and adjust to portuguese. While we don't discuss this I will continue with the translations.

Kindest,
Carol.